Commit - 202211261219
Project complete with examples

Added iomanip library for alignment in console output. Also added a define statement to shorten the length of the cout statements. Also, added printExamples function that prints the examples provided by the instructor along with examples from the book to verify that everything is working correctly.

Added precedent level function to make a determination of the precedence level of each character to be added to the stack